// src/store/accessories/accessories.enum.ts
enum AccessoriesActionTypeEnum {
  GET_ACCESSORIES_PENDING = "GET_ACCESSORIES_PENDING",
  GET_ACCESSORIES_SUCCESS = "GET_ACCESSORIES_SUCCESS",
  GET_ACCESSORIES_FAILED = "GET_ACCESSORIES_FAILED",

  ADD_ACCESSORIES_PENDING = "ADD_ACCESSORIES_PENDING",
  ADD_ACCESSORIES_SUCCESS = "ADD_ACCESSORIES_SUCCESS",
  ADD_ACCESSORIES_FAILED = "ADD_ACCESSORIES_FAILED",

  UPDATE_ACCESSORIES_PENDING = "UPDATE_ACCESSORIES_PENDING",
  UPDATE_ACCESSORIES_SUCCESS = "UPDATE_ACCESSORIES_SUCCESS",
  UPDATE_ACCESSORIES_FAILED = "UPDATE_ACCESSORIES_FAILED",

  DELETE_ACCESSORIES_PENDING = "DELETE_ACCESSORIES_PENDING",
  DELETE_ACCESSORIES_SUCCESS = "DELETE_ACCESSORIES_SUCCESS",
  DELETE_ACCESSORIES_FAILED = "DELETE_ACCESSORIES_FAILED",
}

export default AccessoriesActionTypeEnum;